server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic-settings>=2.9.1",
    "pydantic>=2.11.4",
    "uvicorn>=0.34.2",
    "sqlalchemy>=2.0.41",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.4.0",
    "passlib[bcrypt]>=1.7.4",
    "loguru>=0.7.3",
    "python-multipart>=0.0.20",
    "bcrypt<4.0.0",
]

[tool.mypy]
python_version = "3.13"
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
implicit_optional = false
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
show_error_code_links = true
pretty = true
color_output = true
show_absolute_path = true
plugins = ["returns.contrib.mypy.returns_plugin"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
line-length = 88
indent-width = 4



[tool.ruff.lint]
allowed-confusables = ["−", "ρ", "∗"]
fixable = ["ALL"]
ignore = []
select = ["E4", "E7", "E9", "F", "B", "Q"]
unfixable = []



[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401", "E402", "F403"]
"**/{tests,docs,tools}/*" = ["E402", "F403"]
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.10",
]

[tool.ruff.lint.pydocstyle]
convention = "google"
